refactored even more the contact list component
