+++ thread/src/components/Contacts/Tabs.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Container, Header, Content, Button, Text, Tabs, Tab, ScrollableTab } from 'native-base';
 
-import Contacts from './Contacts';
+import ContactList from './ContactList';
 import RequestsReceived from './RequestsReceived';
 import RequestsSent from './RequestsSent';
 
 const ContactsTabs = () => (
 	<Container>
 		<Tabs renderTabBar={()=> <ScrollableTab />}>
-			<Tab heading="Contacts" children={<Contacts />} />
+			<Tab heading="Contacts" children={<ContactList />} />
 			<Tab heading="Requests Received" children={<RequestsReceived />} />
 			<Tab heading="Requests Sent" children={<RequestsSent />} />
 		</Tabs>
